refactor: implementing abstract api class
This move the general logic from the FuturexAPIClient to a new Abstract class.
